fixed some stack issues
For every function call, ax was PUSHed once and POP'd 2*r times, where unsigned int r = [bp+10] is the radius. This can place the stack pointer (sp regiser) in non-predictable positions. Balanced this by removing the PUSH intruction and the POP. Instead, when ax holds the [bp+10] value, it's saved to [bp-16] via a buffer.